requirements.txt
streamlit
rembg
pillow
onnxruntime
pillow_heif